Added new documentation, improved display functionality. Licensed.
Documentation comments have been added and edited.

The Apache License version 2.0 hase been applied to the project.

The draw instruction has been fixed and should be working according to
the documentation now.  Previously, it wasn't XOR'ing the bits like it
was supposed to, leading to some pixels not being properly displayed.
Also the method for evaluating whether a pixel in the sprit was "on" or
"off" has been fixed.