Another update to the emulation loop, and timer mechanism.
Timers are now internally represented as 16 bit registers that hold
values in milliseconds.  This makes updating them easier and
straightforward.

Also, timing for the ellapsed time between execution of operations
has changed.  Timing now starts AFTER the last instruction has been
executed for the current emulation loop, and ends right before the next
instruction in the following emulation loop.  This is important because
instructions that could potentially take a long time, like LD_Vx_K,
which hangs until a user presses a key, will cause the calculated
ellapsed time quite large, and will cause the emulator to execute a
large number of instrutions in succession.

TLDR: we are only measuring how long it takes the emulation loop to reach
the opcode execution time, and not including the time it takes to execute
said opcodes.