Fairly large commit.  Mostly refactoring of chip8emu.
chip8emu has been refactored so that it contains only the things that deal
directly with the emulation of the instructions.  That means that
keyboard and display functionality have been moved out into their own
files. state for the display and keyboard are maintained in their own
structs, which the main chip8 struct contains.  Naturally the
CMakeLists.txt file has also been changed so that the project still
compiles correctly.